requirements.txt
fastapi
uvicorn
asyncpg
sqlalchemy
python-dotenv
passlib[bcrypt]
supabase